+++ TigGUI/MainWindow.py
@@ -72,7 +72,7 @@ class MainWindow(QMainWindow):
         for icol, col in enumerate(self.ViewModelColumns):
             setattr(self, "Column%s" % col.capitalize(), icol)
         # init GUI
-        self.setWindowTitle(f"Tigger v{TigGUI.__version__}")
+        self.setWindowTitle("Tigger")
         self.setWindowIcon(QIcon(pixmaps.purr_logo.pm()))
         # central widget setup
         self.cw = QWidget(self)